+++ R/phrases.R
@@ -46,6 +46,8 @@ setMethod("phrasetotoken", signature = c("corpus", "ANY"),
               object
           })
 
+setOldClass("tokenizedTexts")
+setClassUnion("textORtokens", members =  c("character", "tokenizedTexts"))
 
 #' @rdname phrasetotoken
 #' @export
@@ -57,7 +59,7 @@ setMethod("phrasetotoken", signature = c("corpus", "ANY"),
 #'          "Some damn good stuff, like the text, she likes that too.")
 #' phrasetotoken(txt, myDict)
 #'
-setMethod("phrasetotoken", signature = c("character", "dictionary"), 
+setMethod("phrasetotoken", signature = c("textORtokens", "dictionary"), 
           function(object, phrases, ...) {
               phraseConcatenator <- phrases@concatenator
               phrasesTmp <- unlist(phrases, use.names = FALSE)
@@ -72,9 +74,11 @@ setClass("collocations", contains = "data.table")
 
 #' @rdname phrasetotoken
 #' @export
-setMethod("phrasetotoken", signature = c("character", "collocations"), 
+setMethod("phrasetotoken", signature = c("textORtokens", "collocations"), 
           function(object, phrases, ...) {
               word1 <- word2 <- word3 <- NULL
+              # sort by word3 so that trigrams will be processed before bigrams
+              data.table::setorder(phrases, -word3, word1)
               # concatenate the words                               
               word123 <- phrases[, list(word1, word2, word3)]
               mwes <- apply(word123, 1, paste, collapse=" ")
@@ -97,10 +101,13 @@ setMethod("phrasetotoken", signature = c("character", "character"),
           function(object, phrases, concatenator = "_", valuetype = c("glob", "regex", "fixed"), 
                    case_insensitive = TRUE, ...) {
               valuetype <- match.arg(valuetype)
-              if (valuetype == "glob" | valuetype == "fixed")
+              if (valuetype == "glob" | valuetype == "fixed") {
                   compoundPhrases <- stringi::stri_replace_all_fixed(phrases, c("*", "?"), 
                                                                      c("[^\\s]*", "[^\\s]"), 
                                                                      vectorize_all = FALSE)
+                  # replace any + symbols that are tokens by escaped \\+ #239
+                  compoundPhrases <- stringi::stri_replace_all_regex(phrases, "(\\s{0,1})\\+(\\s{0,1})", "$1\\\\\\+$2")
+              }
               
               compoundPhrasesList <- strsplit(compoundPhrases, "\\s")
               
@@ -111,5 +118,35 @@ setMethod("phrasetotoken", signature = c("character", "character"),
               }
               object
           })
+
+
+#' @rdname phrasetotoken
+#' @export
+#' @examples 
+#' # on simple text
+#' toks <- tokenize("Simon sez the multi word expression plural is multi word expressions, Simon sez.")
+#' phrases <- c("multi word expression*", "Simon sez")
+#' phrasetotoken(toks, phrases)
+#' 
+setMethod("phrasetotoken", signature = c("tokenizedTexts", "character"), 
+          function(object, phrases, concatenator = "_", valuetype = c("glob", "regex", "fixed"), 
+                   case_insensitive = TRUE, ...) {
+              valuetype <- match.arg(valuetype)
+              
+              # convert any patterns to fixed matches
+              phrasesTok <- tokenize(phrases, what = "fasterword")
+              attr.orig <- attributes(phrasesTok)
+              class.orig <- class(phrasesTok)
+              
+              if (valuetype %in% c("glob", "fixed"))
+                  phrasesTok <- lapply(phrasesTok, glob2rx)
+              phrasesTokFixed <- quanteda:::regexToFixed(object, phrasesTok, case_insensitive = case_insensitive)
+              attributes(phrasesTokFixed) <- attr.orig
+              class(phrasesTokFixed) <- class.orig
+              
+              joinTokens(object, phrasesTokFixed, valuetype = "fixed", case_insensitive = FALSE)
+})
+
+