+++ R/util.R
@@ -1,6 +1,6 @@
-#' @rdname mktemp
-#' make temporary files and directories in a more reasonable way than tempfile()
-#' or tempdir(): here, the filename is different each time you call mktemp()
+# @rdname catm
+# make temporary files and directories in a more reasonable way than tempfile()
+# or tempdir(): here, the filename is different each time you call mktemp()
 mktemp <- function(prefix='tmp.', base_path=NULL, directory=F) {
     #  Create a randomly-named temporary file or directory, sort of like
     #  https://www.mktemp.org/manual.html
@@ -29,9 +29,9 @@ mktemp <- function(prefix='tmp.', base_path=NULL, directory=F) {
 }
 
 
-#' @rdname catm
-#' messages() with some of the same syntax as cat(): takes a sep argument and
-#' does not append a newline by default
+# @rdname catm
+# messages() with some of the same syntax as cat(): takes a sep argument and
+# does not append a newline by default
 catm <- function(..., sep = " ", appendLF = F) {
     message(paste(..., sep = sep), appendLF = appendLF)
 }