+++ R/wordstem.R
@@ -42,7 +42,7 @@ wordstem.character <- function(x, language = "porter") {
 
 # 
 # toks <- unlist(tokenize(toLower(inaugTexts[1:5]), removePunct = TRUE, removeNumbers = TRUE), use.names = FALSE)
-# microbenchmark(wordstem(toks), 
+# microbenchmark::microbenchmark(wordstem(toks), 
 #                wordstemP(toks),
 #                simplify2array(parallel::mclapply(toks, wordstem, language=language)))
 
@@ -63,6 +63,26 @@ wordstem.tokenizedTexts <- function(x, language = "porter") {
     result[which(is.na(x))] <- NA
     attributes(result) <- origAttrs
     result
+}
+
+#' @rdname wordstem
+#' @import stringi 
+#' @export
+#' @examples 
+#' \dontshow{
+#' txt <- c(one = "Eating eater eaters eats ate.",
+#'          two = "Taxing taxes taxed my tax return.")
+#' th <- hashTokens(tokenize(toLower(txt)))
+#' wordstem(th)
+#' attr(wordstem(th), "vocabulary")
+#' }
+wordstem.tokenizedTextsHashed <- function(x, language = "porter") {
+    if (all.equal(attributes(x)$ngrams, 1))
+        vocabulary(x) <- wordstem(vocabulary(x), language = language)
+    else 
+        vocabulary(x) <- wordstem_Ngrams(vocabulary(x), language = language)
+
+    tokens_hashed_recompile(x)
 }